+++ crates/nostr-sdk/CHANGELOG.md
@@ -30,6 +30,7 @@
 - Replace `usize` with `u8` for gossip relay limits
 - Remove `autoconnect` option from `ClientOptions`
 - Change `Client::stream_events*` output to include the `RelayUrl` and `Result`, enabling callers to identify which relay sent the event or if a specific relay encountered an error (https://github.com/rust-nostr/nostr/pull/1156)
+- Change `Client::fetch_events*` to return `Output<Events>` instead of `Events` (TBD)
 
 ## v0.44.1 - 2025/11/09